requirements.txt
django==4.1.4
flake8==6.0.0
pytest-django==4.5.2
six==1.16.0